+++ raphtory-graphql/src/data.rs
@@ -61,7 +61,7 @@ impl Data {
             })
             .collect()
     }
-
+    #[allow(dead_code)]
     // TODO: use this for loading both regular and vectorised graphs
     pub fn generic_load_from_file<T, F>(path: &str, loader: F) -> impl Iterator<Item = T>
     where
