+++ raphtory/src/db/api/mutation/deletion_ops.rs
@@ -9,7 +9,6 @@ use crate::{
             internal::{InternalAdditionOps, InternalDeletionOps},
             TryIntoInputTime,
         },
-        view::StaticGraphViewOps,
     },
 };
