+++ raphtory/src/db/graph/edge.rs
@@ -69,7 +69,7 @@ impl<'graph, G: GraphViewOps<'graph>, GH: GraphViewOps<'graph>> EdgeView<G, GH> 
             edge,
         }
     }
-
+    #[allow(dead_code)]
     fn layer_ids(&self) -> LayerIds {
         self.graph.layer_ids().constrain_from_edge(self.edge)
     }
