+++ raphtory/src/python/graph/views/mod.rs
@@ -1,1 +1,2 @@
 pub mod graph_view;
+mod graph_view_modules;
