+++ raphtory/src/python/packages/vectors.rs
@@ -152,6 +152,7 @@ pub fn into_py_document(
 }
 
 #[cfg(feature = "python")]
+#[allow(dead_code)]
 pub struct PyDocumentTemplate {
     graph_document: Option<String>,
     node_document: Option<String>,
